Refactor log-validator to break up the big main
The main function in log-validator is overly big, so this refactors it in
preparation for adding support for file globs.

Largely it's functionally identical, except for one change:
The 1-per-second ratelimiter is moved to be per-file, so it's
1-per-second-per-file.  I think this more closely aligns with what we'd want,
as we could potentially miss that a file had bad lines in it if it was
overwhelmed by another log file at the same time.
